backend/Calender/calendar-app/requirements.txt
Flask==2.3.3
firebase-admin==6.5.0
python-dotenv==1.0.0
flask-cors==4.0.0